ZEBI

+++ src/main/java/com/miola/smarthotel/dao/EmployeDao.java
@@ -45,7 +45,7 @@ public class EmployeDao implements Dao<Employe>
                             rs.getString("userName"),
                             rs.getString("password"));
                     employe.setIdUser(rs.getInt("idUser"));
-                    employe.setIdEmploye(rs.getLong("idEmploye"));
+                    employe.setIdEmploye(rs.getInt("idEmploye"));
                 }
                 else
                 {
@@ -58,7 +58,7 @@ public class EmployeDao implements Dao<Employe>
                             rs.getString("userName"),
                             rs.getString("password"));
                     employe.setIdUser(rs.getInt("idUser"));
-                    employe.setIdEmploye(rs.getLong("idEmploye"));
+                    employe.setIdEmploye(rs.getInt("idEmploye"));
                 }
             }
             else
@@ -135,7 +135,7 @@ public class EmployeDao implements Dao<Employe>
                             rs.getString("userName"),
                             rs.getString("password"));
                     employe.setIdUser(rs.getInt("idUser")); // c'est l'id de la table user
-                employe.setIdEmploye(rs.getLong("idEmploye"));
+                employe.setIdEmploye(rs.getInt("idEmploye"));
                     employes.add(employe);
             }
             rs.close();
