+++ src/main/java/com/miola/smarthotel/model/Admin.java
@@ -7,7 +7,7 @@ public class Admin extends Employe
         super(prenom, nom, cin, email, telephone, adresse, userName, password);
     }
 
-    public Admin(long idEmploye, String prenom, String nom, String cin, String email, String telephone, String adresse, String userName, String password)
+    public Admin(int idEmploye, String prenom, String nom, String cin, String email, String telephone, String adresse, String userName, String password)
     {
         super(idEmploye, prenom, nom, cin, email, telephone, adresse , userName, password);
     }
