+++ src/main/java/com/miola/smarthotel/model/Receptionniste.java
@@ -7,7 +7,7 @@ public class Receptionniste extends Employe
         super(prenom, nom, cin, email, telephone, adresse, userName, password);
     }
 
-    public Receptionniste(long idEmploye, String prenom, String nom, String cin, String email, String telephone, String adresse, String userName, String password)
+    public Receptionniste(int idEmploye, String prenom, String nom, String cin, String email, String telephone, String adresse, String userName, String password)
     {
         super(idEmploye, prenom, nom, cin, email, telephone, adresse, userName, password);
     }
